Move config generator to 'cli' package
